using ThunkDispatch type for AppDispatch, moving AppDispatch definition into the store, and adding Thunk middleware
