+++ src/actions/index.ts
@@ -1,6 +1,3 @@
-import { Dispatch } from 'redux';
 import { CartActionTypes } from './cart';
-
-export type AppDispatch = Dispatch<CartActionTypes>;
 
 export type AppActions = CartActionTypes;
