+++ src/components/ProductCard.tsx
@@ -3,7 +3,7 @@ import {
   Card,
   Button,
 } from 'react-bootstrap';
-import { AppDispatch } from '../actions';
+import { AppDispatch } from '../store';
 import { addToCart } from '../actions/cart';
 import { Product } from '../types/product';
 import { useDispatch } from 'react-redux';
